+++ craco.config.js
@@ -1,0 +1,17 @@
+const CracoLessPlugin = require("craco-less");
+
+module.exports = {
+  plugins: [
+    {
+      plugin: CracoLessPlugin,
+      options: {
+        lessLoaderOptions: {
+          lessOptions: {
+            javascriptEnabled: true,
+            modifyVars: { "@primary-color": "#018a9e" },
+          },
+        },
+      },
+    },
+  ],
+};
